pyproject.toml
[tool.poetry]
name = "fastapi-lms"
version = "0.1.0"
description = "This is an entry for fastapi"
authors = ["Stephen Gbolagade <gbolabyte222@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.13"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
sqlmodel = "^0.0.22"
sqlalchemy-utils = "^0.41.2"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.37"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
